requirements.txt
django==3.1
django-crispy-forms==1.9.2
pytube3